Text-stream buffer perf, suppress-prompts setting, and shared window-view scaffold
ComposeTextStream: update the displayed lines incrementally instead of re-filtering the whole buffer on every line; ArrayDeque line buffers and an offset-based persistent view so per-newline work at capacity is O(log n); an identity-equals snapshot so StateFlow/Compose dedup is O(1).

Adds a 'suppress prompts' client setting that filters prompt lines from the displayed stream, with desktop and mobile settings toggles.

Extracts a shared WindowViewScaffold so desktop and mobile window rendering go through one implementation, with the platform-specific pieces supplied as slots.
