+++ compose/src/jvmMain/kotlin/warlockfe/warlock3/compose/desktop/ui/settings/DesktopGeneralSettingsView.kt
@@ -144,6 +144,21 @@ fun DesktopGeneralSettingsView(
 
             Spacer(Modifier.height(16.dp))
 
+            val suppressPrompts by clientSettingRepository
+                .observeSuppressPrompts()
+                .collectAsState(initial = false)
+            WarlockCheckboxRow(
+                checked = suppressPrompts,
+                onCheckedChange = {
+                    scope.launch {
+                        clientSettingRepository.putSuppressPrompts(it)
+                    }
+                },
+                text = "Hide prompts",
+            )
+
+            Spacer(Modifier.height(16.dp))
+
             val autoConnectLastConnection by clientSettingRepository
                 .observeAutoConnectLastConnection()
                 .collectAsState(initial = false)
